+++ docs/user_guides/02_quickstart.md
@@ -16,12 +16,13 @@ Every FabricPC workflow follows this pattern:
 
 ```python
 from jax_setup import set_jax_flags_before_importing_jax
+
 set_jax_flags_before_importing_jax(jax_platforms="cuda")
 
 import jax
 from fabricpc.nodes import Linear, IdentityNode
 from fabricpc.builder import Edge, TaskMap, graph
-from fabricpc.graph import initialize_params
+from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import SigmoidActivation, SoftmaxActivation
 from fabricpc.core.energy import CrossEntropyEnergy
 from fabricpc.core.inference import InferenceSGD
