+++ examples/PC_backprop_compare.py
@@ -33,7 +33,7 @@ import importlib.util
 
 from fabricpc.nodes import Linear
 from fabricpc.builder import Edge, TaskMap, graph
-from fabricpc.graph import initialize_params
+from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import (
     SigmoidActivation,
     SoftmaxActivation,
