+++ examples/custom_node.py
@@ -25,7 +25,7 @@ from fabricpc.utils.data.dataloader import MnistLoader
 from fabricpc.nodes import Linear
 from fabricpc.nodes.base import NodeBase, SlotSpec
 from fabricpc.builder import Edge, TaskMap, graph
-from fabricpc.graph import initialize_params
+from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import (
     IdentityActivation,
     ReLUActivation,
