+++ examples/jpc_fc_resnet_compare.py
@@ -105,7 +105,7 @@ import optax
 from fabricpc.nodes import IdentityNode
 from fabricpc.nodes.base import NodeBase, SlotSpec
 from fabricpc.builder import Edge, TaskMap, graph
-from fabricpc.graph import initialize_params
+from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import IdentityActivation, ReLUActivation
 from fabricpc.core.energy import GaussianEnergy
 from fabricpc.core.inference import InferenceSGDNormClip
@@ -604,7 +604,7 @@ def probe_variance(params, structure, test_loader, rng_key):
     Prints a table showing z_mu variance, latent_grad variance, and
     z_latent variance at each layer, revealing signal propagation issues.
     """
-    from fabricpc.graph.state_initializer import initialize_graph_state
+    from fabricpc.graph_initialization.state_initializer import initialize_graph_state
     from fabricpc.core.inference import run_inference
 
     # Get one batch
