+++ examples/mnist_aim_tracking.py
@@ -26,7 +26,7 @@ from fabricpc.utils.data.dataloader import MnistLoader
 
 from fabricpc.nodes import Linear, IdentityNode
 from fabricpc.builder import Edge, TaskMap, graph
-from fabricpc.graph import initialize_params, FeedforwardStateInit
+from fabricpc.graph_initialization import initialize_params, FeedforwardStateInit
 from fabricpc.core.activations import (
     IdentityActivation,
     SigmoidActivation,
