+++ examples/mnist_multi_gpu.py
@@ -20,7 +20,7 @@ import time
 
 from fabricpc.nodes import Linear
 from fabricpc.builder import Edge, TaskMap, graph
-from fabricpc.graph import initialize_params
+from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import (
     IdentityActivation,
     SigmoidActivation,
