+++ examples/resnet18_cifar10_demo.py
@@ -57,7 +57,7 @@ from fabricpc.nodes.base import NodeBase, SlotSpec
 from fabricpc.core.types import NodeParams
 from fabricpc.builder import Edge, TaskMap, graph
 from fabricpc.core import InferenceSGDNormClip
-from fabricpc.graph import initialize_params
+from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import (
     IdentityActivation,
     ReLUActivation,
