+++ examples/storkey_hopfield_recall.py
@@ -37,8 +37,8 @@ import optax
 
 from fabricpc.nodes import IdentityNode, StorkeyHopfield
 from fabricpc.builder import Edge, TaskMap, graph
-from fabricpc.graph import initialize_params
-from fabricpc.graph.state_initializer import initialize_graph_state
+from fabricpc.graph_initialization import initialize_params
+from fabricpc.graph_initialization.state_initializer import initialize_graph_state
 from fabricpc.core.inference import InferenceSGD, run_inference
 from fabricpc.core.initializers import NormalInitializer
 from fabricpc.training import train_pcn
