+++ fabricpc/training/train_backprop.py
@@ -23,7 +23,7 @@ import optax
 
 from fabricpc.core import GraphState
 from fabricpc.core.types import GraphParams, GraphStructure
-from fabricpc.graph.state_initializer import initialize_graph_state
+from fabricpc.graph_initialization.state_initializer import initialize_graph_state
 from fabricpc.training.train_autoregressive import create_causal_mask, compute_loss
 
 
@@ -34,7 +34,7 @@ def validate_feedforward_init(structure: GraphStructure):
     Raises:
         ValueError if incompatible.
     """
-    from fabricpc.graph.state_initializer import FeedforwardStateInit
+    from fabricpc.graph_initialization.state_initializer import FeedforwardStateInit
 
     init = structure.config["graph_state_initializer"]
     if not isinstance(init, FeedforwardStateInit):
