+++ tests/test_energy.py
@@ -19,7 +19,7 @@ from fabricpc.core.energy import (
 )
 from fabricpc.nodes import Linear
 from fabricpc.builder import Edge, TaskMap, graph
-from fabricpc.graph import initialize_params
+from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.inference import InferenceSGD
