+++ tests/test_fabricpc.py
@@ -12,9 +12,9 @@ import jax
 import jax.numpy as jnp
 
 from fabricpc.core.types import NodeState, NodeParams, GraphState
-from fabricpc.graph.graph_net import compute_local_weight_gradients
-from fabricpc.graph import initialize_params
-from fabricpc.graph.state_initializer import initialize_graph_state
+from fabricpc.graph_initialization.graph_net import compute_local_weight_gradients
+from fabricpc.graph_initialization import initialize_params
+from fabricpc.graph_initialization.state_initializer import initialize_graph_state
 from fabricpc.core.inference import InferenceSGD
 import optax
 from fabricpc.training import train_step
