+++ tests/test_multi_gpu.py
@@ -15,7 +15,7 @@ import optax
 
 from fabricpc.nodes import Linear
 from fabricpc.builder import Edge, TaskMap, graph
-from fabricpc.graph import initialize_params
+from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import IdentityActivation, ReLUActivation
 from fabricpc.core.initializers import XavierInitializer
 from fabricpc.core.inference import InferenceSGD
