+++ tests/test_ndim_shapes.py
@@ -12,8 +12,8 @@ import jax.numpy as jnp
 
 from fabricpc.nodes import Linear
 from fabricpc.builder import Edge, TaskMap, graph
-from fabricpc.graph import initialize_params
-from fabricpc.graph.state_initializer import initialize_graph_state
+from fabricpc.graph_initialization import initialize_params
+from fabricpc.graph_initialization.state_initializer import initialize_graph_state
 from fabricpc.core.inference import InferenceSGD
 import optax
 from fabricpc.training import train_step
