+++ tests/test_optimizers.py
@@ -15,7 +15,7 @@ from fabricpc.training.optimizers import (
 from fabricpc.training import train_step
 from fabricpc.nodes import Linear
 from fabricpc.builder import Edge, TaskMap, graph
-from fabricpc.graph import initialize_params
+from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import SigmoidActivation
 from fabricpc.core.inference import InferenceSGD
