+++ tests/test_train_backprop.py
@@ -10,10 +10,10 @@ import optax
 
 from fabricpc.nodes import Linear
 from fabricpc.builder import Edge, TaskMap, graph
-from fabricpc.graph import initialize_params
+from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import ReLUActivation, SoftmaxActivation
 from fabricpc.core.inference import InferenceSGD
-from fabricpc.graph.state_initializer import GlobalStateInit
+from fabricpc.graph_initialization.state_initializer import GlobalStateInit
 from fabricpc.training.train_backprop import (
     train_backprop,
     evaluate_backprop,
